cloud-function-on-img-upload/requirements.txt
google-cloud-vision==2.7.1
google-cloud-firestore==2.0.2
